Add ability to configure CidrIP of security group rules
This adds .spec.securityGroup.*Rules.CidrIp so that the created security
group rules can optionally allow a target IPv4 CIDR range instead of
only being allowed for the cluster's own master and worker security
groups. If a CIDR range is specified, it will take precendence and be
used, while the cluster's own master and worker security groups act as a
fallback default.

Signed-off-by: Michael Shen <mshen@redhat.com>