Handle edge cases when replacing v1alpha1 VpcEndpoints with v1alpha2
While migrating from v1alpha1 to v1alpha2, it was noticed that AVO
reconciles away all the subnets of the VPC Endpoint in AWS, then
eventually fails when executing generateRoute53Record() as there are no
subnets attached to the VPC Endpoint, so it does not have any DNS names.

This commit handles the edge case by allowing that function to return
without error when the VpcEnpoint CR is deleting and adds additional
safeguards to ensure .status.hostedZoneId is populated before attempting
to delete Route53 Record Sets.

Signed-off-by: Michael Shen <mshen@redhat.com>

Co-authored-by: Antony Natale <anatale@redhat.com>