Initial VpcEndpointTemplate CRD and Controller
This PR creates a new CRD, `VpcEndpointTemplate` and an associated
controller. A `VpcEndpointTemplate` behaves like a DaemonSet does to a
Pod. In its current iteration, the controller responds to events about
`VpcEndpointTemplate` CRs as well as `HostedControlPlane` CRs, with the
outcome of a reconcile ensuring that each `HostedControlPlane` has one
`VpcEndpoint` matching the expected template.

Signed-off-by: Michael Shen <mshen@redhat.com>

Co-authored-by: Alice Hubenko <ahubenko@redhat.com>