Add a new .spec.serviceNameRef
This is a more generic form of .spec.serviceName that will allow
VpcEndpoint resources to be templated out. If were were designing this
API from scratch again, .spec.serviceName would be removed, but we
cannot make a breaking change at this point. The VPC Endpoint Service
Name will instead be stored in .status.vpcEndpointServiceName instead.

Signed-off-by: Michael Shen <mshen@redhat.com>