Store cluster infra id in .status.infraId
This commit fixes a race condition during VpcEndpoint deletion when the
backing CustomResources which holds the infra id is deleted before the
VpcEndpoint. By storing the infra id in the VpcEndpoint's own status
field, the deletion of these objects can continue regardless of whether
the infrastructure or hostedcontrolplane CRs are deleted first.

Signed-off-by: Michael Shen <mshen@redhat.com>