Fix rejected VPCE preventing cleanup of Route53
This commit fixes a bug where a rejected VPC Endpoint would cause the
generation of a Route53 Record to fail, which was halting the deletion
of a VpcEndpoint. This is fixed by not trying to generate a Route53
record and depending on the status field.

If we own a hosted zone, then just cleanup all records during deletion.

Signed-off-by: Michael Shen <mshen@redhat.com>